*** empty log message ***
git-svn-id: svn://svn.macaulay2.com/Macaulay2/trunk/M2@3355 a8012f80-5f13-0410-8dc0-eb77bd940c0d

+++ Macaulay2/e/x_gb.cpp
@@ -286,7 +286,7 @@ IM2_GB_to_string(Computation *C)
   return o.to_string();
 }
 
-int IM2_GB_hash(const Computation *C)
+unsigned long IM2_GB_hash(const Computation *C)
 {
   return C->get_hash_value();
 }
